Polish apache/dubbo#6322 : [Enhancement] Fix the issues of test-cases after refactoring
